requirements.txt
channels==4.1.0
daphne==4.1.2
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
mysqlclient==2.2.4
openai==1.35.10
Pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
